move type convertsion rule into optimizer dir
