+++ src/query/src/query_engine/state.rs
@@ -33,7 +33,7 @@ use datafusion_optimizer::analyzer::Analyzer;
 use promql::extension_plan::PromExtensionPlanner;
 
 use crate::dist_plan::{DistExtensionPlanner, DistPlannerAnalyzer};
-use crate::optimizer::TypeConversionRule;
+use crate::optimizer::type_conversion::TypeConversionRule;
 use crate::query_engine::options::QueryOptions;
 
 /// Query engine global state
